requirements.txt
aioconsole==0.7.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
cattrs==23.2.3
certifi==2024.7.4
cfgv==3.4.0
distlib==0.3.8
distro==1.9.0
filelock==3.15.4
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
identify==2.6.0
idna==3.8
jiter==0.5.0
nodeenv==1.9.1
numpy==2.0.1
openai==1.41.0
platformdirs==4.2.2
pre-commit==3.8.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
PyYAML==6.0.2
ruff==0.6.1
sniffio==1.3.1
tabulate==0.9.0
termcolor==2.4.0
tqdm==4.66.5
typing_extensions==4.12.2
virtualenv==20.26.3
